Enable IP customization using DHCP
This changes updates the `CustomizationSpec` applied to the cloned VM so
that it's IP address can be generated automatically using DHCP if a
fixed IPv4 address is not specified. To make this possible the
validations on the guest customization settings had to be adjusted to
allow `ip_address`, `subnet_mask`, and `gateway` to be optional instead
of required. Now if the `ip_address` is omitted the VM's customization
spec will use the `CustomizationDhcpIpGenerator` when setting up the
customized ip settings.

Signed-off-by: Keith Walters <kwalters@taphere.com>